+++ Pulp/Core/win32/OsFile.h
@@ -16,11 +16,11 @@ struct OsFile
 	size_t read(void* buffer, size_t length);
 	size_t write(const void* buffer, size_t length);
 
-	void seek(size_t position, IFileSys::SeekMode::Enum origin);
+	void seek(int64_t position, IFileSys::SeekMode::Enum origin);
 
-	size_t tell(void) const;
-	size_t remainingBytes(void) const;
-	void setSize(size_t numBytes);
+	uint64_t tell(void) const;
+	uint64_t remainingBytes(void) const;
+	void setSize(int64_t numBytes);
 
 	bool valid(void) const;
