+++ Pulp/RenderSystems/Common/Textures/TextureLoaderDDS.cpp
@@ -986,7 +986,7 @@ namespace DDS
 		}
 
 #if X_DEBUG == 1
-		size_t left = file->remainingBytes();
+		uint64_t left = file->remainingBytes();
 		X_WARNING_IF(left > 0, "DDSLoader", "potential read fail, bytes left in file: %i", left);
 #endif
