+++ Pulp/RenderSystems/Common/Textures/TextureLoaderTGA.cpp
@@ -151,7 +151,7 @@ XTextureFile* XTexLoaderTGA::loadTexture(core::XFile* file)
 	}
 
 #if X_DEBUG == 1
-	size_t left = file->remainingBytes();
+	uint64_t left = file->remainingBytes();
 	X_WARNING_IF(left > 0, "TextureTGA", "potential read fail, bytes left in file: %i", left);
 #endif
